// e2e/pages/components/navbar/subOptions.d.ts
export  enum SWITCHTO{
    Alerts,
    Windows,
    Frames
}

export enum WIDGETS{
    Accordion,
    AutoComplete,
    Datepicker,
    Slider,
}

export enum INTERACTIONS{
    DragandDrop="Drag and Drop",
    Selectable,
    Resizable


}

export enum VIDEO{
    Youtube,
    Vimeo
}

export enum WYSIWYG{
    TinyMCE,
    CKEditor,
    SummerNote,
    CodeMirror
}

export enum MORE{
    Charts,
    DynamicData="Dynamic Data",
    FileDownload="File Download",
    FileUpload="File Upload",
    JQueryProgressBar="JQuery ProgressBar",
    Loader,
    Modals,
    ProgressBar
}